fix: respect custom coroutine scope when calling blocking functions in a data fetcher
When running blocking function data fetcher it does not respect coroutine scope provided in the GraphQL context.

In my project we're creating custom a coroutine scope in order to propagate MDC and OTEL contexts in a coroutine like this:

```
CoroutineScope(MDCContext() + Dispatchers.IO + Context.current().asContextElement())
```

(here `Context` is an instance of OTEL context).

This works well in suspended functions since they are using the context however normal functions are not using this context and therefore none of the OTEL context (trace id etc) is propagated.